Kiali-589 Remove graph filter toolbar/toggle buttons and replace with popover/checkboxes.
